requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
confluent-kafka==1.8.2
decorator==5.1.1
delta-spark==1.0.0
executing==0.8.3
Faker==13.3.3
faker-commerce==1.0.3
findspark==2.0.1
idna==3.3
importlib-metadata==4.11.3
ipython==8.2.0
jedi==0.18.1
kafka-python==2.0.2
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9
Pygments==2.11.2
pyspark==3.1.3
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
stack-data==0.2.0
traitlets==5.1.1
urllib3==1.26.9
wcwidth==0.2.5
zipp==3.8.0
